Combine user profile pages
